lws_struct: clean warning on windows

+++ lib/misc/lws-struct-lejp.c
@@ -52,7 +52,7 @@ lws_struct_schema_only_lejp_cb(struct lejp_ctx *ctx, char reason)
 
 		while (n--) {
 			const lws_struct_map_t *child = map->child_map;
-			int m, child_members = map->child_map_size;
+			int m, child_members = (int)map->child_map_size;
 
 			for (m = 0; m < child_members; m++) {
 				if (!strcmp(ctx->path, child->colname)) {
